feat: add xp login

+++ ZenGarden/ZenGarden.Domain/Entities/UserXpLog.cs
@@ -8,7 +8,7 @@ public class UserXpLog
 
     public int UserId { get; init; }
     public XpSourceType XpSource { get; init; }
-    public int XpAmount { get; init; }
+    public double XpAmount { get; init; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
     public Users? User { get; init; }
